server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
SQLAlchemy==2.0.23
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0

# File handling and security
python-magic==0.4.27
python-magic-bin==0.4.14

# Additional utilities
uuid==1.30
python-dateutil==2.8.2

# AI and Machine Learning dependencies
ultralytics>=8.0.0
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.3.0
Pillow>=8.0.0
